many modyfications in search for more speed

+++ util/ceg/ce_back/obj_back/con2.c
@@ -6,11 +6,11 @@ TWO_BYTES w;
 {
 	if ((_data_cnt -= 2) < 0) mem_data();
 #ifdef BYTES_REVERSED
-	*data++ = ( unsigned short)w>>8;
+	*data++ = w>>8;
 	*data++ = w;
 #else
 	*data++ = w;
-	*data++ = ( unsigned short)w>>8;
+	*data++ = w>>8;
 #endif
 }
