+++ util/ceg/ce_back/obj_back/text2.c
@@ -6,10 +6,10 @@ TWO_BYTES w;
 {
 	if ((_text_cnt -= 2) < 0) mem_text();
 #ifdef BYTES_REVERSED
-	*text++ = ( unsigned short)w>>8;
+	*text++ = w>>8;
 	*text++ = w;
 #else
 	*text++ = w;
-	*text++ = ( unsigned short)w>>8;
+	*text++ = w>>8;
 #endif
 }
